storage: remove repeated retry logic that causes indefinite waits
The internals from google.golang.org/api/gensupport already
perform retries on uploading. Adding retries in this package
meant that if the gensupport retries were exhausted, we'd
lose resumable upload state yet retry on a non-rewound
media body, and make a fresh requests, which would exacerbate
the problem that started with a 429 or so and eventually turn
it into a 503 Backend API error, such as:

    TransferChunk: 503 Service Unavailable
    Body in error: mismatched_content_start
    (Invalid request. According to the Content-Range header,
    the upload offset is 812384256 byte(s), which exceeds
    already uploaded size of 0 byte(s).)
    [bytes 812384256-1082166531/1082166532]

and then cause more retries with a fresh request, ad infinitum.

This change is on an experimental branch to be shipped out to
the issue reporter.

Updates #1507.